refactor: 매니저에 의한 운동 참여 취소 메서드 참여 서비스로 이동
