+++ src/main/java/umc/cockple/demo/domain/exercise/service/ExerciseParticipationService.java
@@ -6,11 +6,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import umc.cockple.demo.domain.exercise.converter.ExerciseConverter;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
+import umc.cockple.demo.domain.exercise.domain.Guest;
 import umc.cockple.demo.domain.exercise.dto.ExerciseCancelDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseJoinDTO;
 import umc.cockple.demo.domain.exercise.exception.ExerciseErrorCode;
 import umc.cockple.demo.domain.exercise.exception.ExerciseException;
 import umc.cockple.demo.domain.exercise.repository.ExerciseRepository;
+import umc.cockple.demo.domain.exercise.repository.GuestRepository;
 import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.domain.member.domain.MemberExercise;
 import umc.cockple.demo.domain.member.repository.MemberExerciseRepository;
@@ -29,6 +31,7 @@ public class ExerciseParticipationService {
     private final MemberRepository memberRepository;
     private final MemberPartyRepository memberPartyRepository;
     private final MemberExerciseRepository memberExerciseRepository;
+    private final GuestRepository guestRepository;
 
     private final ExerciseValidator exerciseValidator;
 
@@ -78,6 +81,23 @@ public class ExerciseParticipationService {
         return exerciseConverter.toCancelResponse(exercise, member);
     }
 
+    public ExerciseCancelDTO.Response cancelParticipationByManager(
+            Long exerciseId, Long participantId, Long memberId, ExerciseCancelDTO.ByManagerRequest request) {
+
+        log.info("매니저에 의한 운동 참여 취소 시작 - exerciseId: {}, participantId: {}, memberId: {}", exerciseId, participantId, memberId);
+
+        Exercise exercise = findExerciseOrThrow(exerciseId);
+        Member manager = findMemberOrThrow(memberId);
+        exerciseValidator.validateCancelCommonParticipationByManager(exercise, manager);
+
+        ExerciseCancelDTO.Response response = executeParticipantCancellation(exercise, participantId, request);
+
+        log.info("매니저에 의한 운동 참여 취소 완료 - exerciseId: {}, participantId: {}, 현재 참여자 수: {}",
+                exerciseId, participantId, exercise.getNowCapacity());
+
+        return response;
+    }
+
     // ========== 비즈니스 메서드 ============
 
     private boolean isPartyMember(Exercise exercise, Member member) {
@@ -85,11 +105,53 @@ public class ExerciseParticipationService {
         return memberPartyRepository.existsByPartyAndMember(party, member);
     }
 
+    private ExerciseCancelDTO.Response executeParticipantCancellation(Exercise exercise, Long participantId, ExerciseCancelDTO.ByManagerRequest request) {
+        if(request.isGuest()){
+            log.info("게스트 참여 취소 실행 - participantId: {}", participantId);
+            return cancelGuestParticipation(exercise, participantId);
+        }
+
+        log.info("멤버 참여 취소 실행 - participantId: {}", participantId);
+        return cancelMemberParticipation(exercise, participantId);
+    }
+
+    private ExerciseCancelDTO.Response cancelGuestParticipation(Exercise exercise, Long participantId) {
+        Guest guest = findGuestOrThrow(participantId);
+        exerciseValidator.validateCancelGuestParticipationByManager(guest, exercise);
+
+        exercise.removeGuest(guest);
+
+        guestRepository.delete(guest);
+
+        exerciseRepository.save(exercise);
+
+        return exerciseConverter.toCancelResponse(exercise, guest);
+    }
+
+    private ExerciseCancelDTO.Response cancelMemberParticipation(Exercise exercise, Long participantId) {
+        Member participant = findMemberOrThrow(participantId);
+        MemberExercise memberExercise = findMemberExerciseOrThrow(exercise, participant);
+
+        exercise.removeParticipation(memberExercise);
+        participant.removeParticipation(memberExercise);
+
+        memberExerciseRepository.delete(memberExercise);
+
+        exerciseRepository.save(exercise);
+
+        return exerciseConverter.toCancelResponse(exercise, participant);
+    }
+
     // ========== 조회 메서드 ============
 
     private Exercise findExerciseOrThrow(Long exerciseId) {
         return exerciseRepository.findById(exerciseId)
                 .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.EXERCISE_NOT_FOUND));
+    }
+
+    private Guest findGuestOrThrow(Long guestId) {
+        return guestRepository.findById(guestId)
+                .orElseThrow(() -> new ExerciseException(ExerciseErrorCode.GUEST_NOT_FOUND));
     }
 
     private Exercise findExerciseWithPartyLevelOrThrow(Long exerciseId) {
